[#9]テクスチャ周りに新ファイルを追加
