+++ 秋制作 ProjectA ver_α/テンプレート/Src/Scene/Scene.h
@@ -14,6 +14,7 @@ enum SceneId
 {
 	TitleScene,			//!< タイトルシーン
 	GameScene,
+	EnemyScene,
 	CenterScene,
 	LeftScene,
 	RightScene,
