+++ 秋制作 ProjectA ver_α/テンプレート/Src/Texture/EnemyTex.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "GameSceneTex.h"
+
+enum EnemyCategoryTextureList
+{
+	FredyTex,
+	EnemyBoonieTex,
+	EnemyTextureMax,
+};
+
+#define TEXTURE_CATEGORY_ENEMY (SceneId::EnemyScene)				//!< ゲーム本編カテゴリー
