+++ 秋制作 ProjectA ver_α/テンプレート/Src/Texture/ResultSceneTex.h
@@ -1,0 +1,10 @@
+#pragma once
+
+enum ClearCategoryTextureList
+{
+	ClearBgTex,				//!< ”wŒi
+	ClearGameOverTex,
+	ClearTextureMax,			//!< ƒŠƒXƒgÅ‘å”
+};
+
+#define TEXTURE_CATEGORY_CLEAR (SceneId::ClearScene)				//!< ƒwƒ‹ƒvƒJƒeƒSƒŠ[
