+++ 秋制作 ProjectA ver_α/テンプレート/Src/Texture/Texture.h
@@ -7,128 +7,11 @@
 
 #include "../Scene/Scene.h"
 #include "../Engine/Graphics.h"
+#include "EnemyTex.h"
+#include "TitleSceneTex.h"
+#include "GameSceneTex.h"
+#include "ResultSceneTex.h"
 
-/** @brief タイトル用テクスチャのリスト */
-enum TitleCategoryTextureList
-{
-	TitleBgTex,				//!< 背景
-	TitleLogoTex,		
-	TitleNewGameUITex,
-	TitleContinueUITex,
-
-	TitleFreeTex,
-
-	TitleTextureMax			//!< リスト最大数
-};
-
-/** @brief ゲーム本編用テクスチャのリスト */
-enum GameCategoryTextureList
-{
-	GameMaskTex,				//!< 背景
-
-	//item
-	GameCandleBigTex,
-	GameCandleSmallTex,
-	GameCandleStandTex,
-	GameCandleEffectTex,
-	GameFireBigTex,
-	GameFireSmallTex,
-	GameCrystalTex,
-	GameMusicBoxTex,
-
-	GameLeftShojiFrontTex,
-	GameLeftShojiBackTex,
-	GameRightShojiFrontTex,
-	GameRightShojiBackTex,
-
-	GameFreeTex,
-
-	GameTextureMax,			//!< リスト最大数
-};
-
-enum CenterCategoryTextureList {
-	GameCenterBgTex,
-
-	GameCenterFreddy,
-	GameCenterPuppet,
-
-	GameCircle1Tex,
-	GameCircle2Tex,
-	GameCircle3Tex,
-	GameCircle4Tex,
-	GameCircle5Tex,
-	GameCircle6Tex,
-	GameCircle7Tex,
-	GameCircle8Tex,
-	
-
-
-	GameCenterTextureMax,			//!< リスト最大数
-};
-
-enum RightCategoryTextureList {
-	GameRightBgTex,
-
-	GameRightChicaTex,
-
-	GameRightTextureMax,			//!< リスト最大数
-};
-
-enum LeftCategoryTextureList {
-	GameLeftBgTex,
-	GameLeftTextureMax,			//!< リスト最大数
-};
-
-enum MonitorCategoryTextureList {
-	GameMonitorSpownBgTex,
-	GameMonitorLeftDuctBgTex,
-	GameMonitorRightDuctBgTex,
-
-	GameMonitorChildRoomBgTex,
-	GameMonitorResptionRoomBgTex,
-	GameMonitorStoreRoomBgTex,
-	
-	GameMonitorSpownUITex,
-	GameMonitorLeftDuctUITex,
-	GameMonitorRightDuctUITex,
-	
-	GameMonitorOnSpownUITex,
-	GameMonitorOnLeftDuctUITex,
-	GameMonitorOnRightDuctUITex,
-
-	GameMonitorPlayerRoomUITex,
-
-	GameMonitorFreddy,
-	GameMonitorChicaTex,
-	GameMonitorPuppetTex,
-
-	GameMonitorMapTex,
-	GameMonitorButtom1Tex,
-	GameMonitorButtom2Tex,
-	GameMonitorButtom3Tex,
-	GameMonitorButtom4Tex,
-	GameMonitorButtom5Tex,
-	GameMonitorButtom6Tex,
-
-
-	GameMonitorTextureMax,			//!< リスト最大数
-};
-
-/** @brief ヘルプ用テクスチャリスト */
-enum ClearCategoryTextureList
-{
-	ClearBgTex,				//!< 背景
-	ClearGameOverTex,
-	ClearTextureMax,			//!< リスト最大数
-};
-
-#define TEXTURE_CATEGORY_TITLE (SceneId::TitleScene)			//!< タイトルカテゴリー
-#define TEXTURE_CATEGORY_GAME (SceneId::GameScene)				//!< ゲーム本編カテゴリー
-#define TEXTURE_CATEGORY_CENTER (SceneId::CenterScene)				//!< ゲーム本編カテゴリー
-#define TEXTURE_CATEGORY_LEFT (SceneId::LeftScene)				//!< ゲーム本編カテゴリー
-#define TEXTURE_CATEGORY_RIGHT (SceneId::RightScene)				//!< ゲーム本編カテゴリー
-#define TEXTURE_CATEGORY_MONITOR (SceneId::MonitorScene)				//!< ゲーム本編カテゴリー
-#define TEXTURE_CATEGORY_CLEAR (SceneId::ClearScene)				//!< ヘルプカテゴリー
 #define MAX_TEXTURE_CATEGORY (SceneId::SceneIdMax)				//!< カテゴリー最大
 
 /**
