+++ 秋制作 ProjectA ver_α/テンプレート/テンプレート.vcxproj.filters
@@ -298,6 +298,18 @@
     <ClInclude Include="Src\Object\UI\UIID.h">
       <Filter>ソース ファイル\Object\UI</Filter>
     </ClInclude>
+    <ClInclude Include="Src\Texture\EnemyTex.h">
+      <Filter>ソース ファイル\Texture</Filter>
+    </ClInclude>
+    <ClInclude Include="Src\Texture\GameSceneTex.h">
+      <Filter>ソース ファイル\Texture</Filter>
+    </ClInclude>
+    <ClInclude Include="Src\Texture\ResultSceneTex.h">
+      <Filter>ソース ファイル\Texture</Filter>
+    </ClInclude>
+    <ClInclude Include="Src\Texture\TitleSceneTex.h">
+      <Filter>ソース ファイル\Texture</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Image Include="Res\Document\マップ図.jpg.png">
